[BE] RENAME, REFACTOR: 기존의 파일들 올바른 패키지로 수정하고, User 어노테이션 User 패키지에 추가 및 이외의 파일들 리팩터링
