+++ backend/src/main/java/org/ftclub/cabinet/auth/domain/ApiUriBuilder.java
@@ -1,4 +1,4 @@
-package org.ftclub.cabinet.auth;
+package org.ftclub.cabinet.auth.domain;
 
 import org.springframework.stereotype.Component;
