+++ backend/src/main/java/org/ftclub/cabinet/auth/domain/AuthAspect.java
@@ -1,4 +1,4 @@
-package org.ftclub.cabinet.auth;
+package org.ftclub.cabinet.auth.domain;
 
 import javax.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
