--- backend/src/main/java/org/ftclub/cabinet/auth/repository/AuthRepository.java
@@ -1,11 +1,0 @@
-package org.ftclub.cabinet.auth.repository;
-
-import org.ftclub.cabinet.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AuthRepository extends JpaRepository<User, Long> {
-
-    User findUserByName(String name);
-}
