+++ backend/src/main/java/org/ftclub/cabinet/cabinet/controller/AdminCabinetController.java
@@ -4,8 +4,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
-import org.ftclub.cabinet.auth.AuthGuard;
-import org.ftclub.cabinet.auth.AuthGuard.Level;
+import org.ftclub.cabinet.auth.domain.AuthGuard;
+import org.ftclub.cabinet.auth.domain.AuthGuard.Level;
 import org.ftclub.cabinet.cabinet.domain.CabinetStatus;
 import org.ftclub.cabinet.cabinet.domain.LentType;
 import org.ftclub.cabinet.cabinet.service.CabinetFacadeService;
