+++ backend/src/main/java/org/ftclub/cabinet/cabinet/controller/CabinetController.java
@@ -3,8 +3,8 @@ package org.ftclub.cabinet.cabinet.controller;
 import java.util.HashMap;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.ftclub.cabinet.auth.AuthGuard;
-import org.ftclub.cabinet.auth.AuthGuard.Level;
+import org.ftclub.cabinet.auth.domain.AuthGuard;
+import org.ftclub.cabinet.auth.domain.AuthGuard.Level;
 import org.ftclub.cabinet.cabinet.service.CabinetFacadeService;
 import org.ftclub.cabinet.dto.BuildingFloorsDto;
 import org.ftclub.cabinet.dto.CabinetDto;
