+++ backend/src/test/java/org/ftclub/cabinet/user/controller/AdminUserControllerTestSession.java
@@ -18,7 +18,7 @@ import org.springframework.test.web.servlet.MockMvc;
 @SpringBootTest
 @AutoConfigureMockMvc
 @Transactional
-class AdminUserControllerTest {
+class AdminUserControllerTestSession {
 
 	@Autowired
 	MockMvc mockMvc;
